requirements.txt
# Shortcut Bridge - Python Dependencies
# Install: pip install -r requirements.txt

# Core server
flask>=2.3.0
flask-cors>=4.0.0

# Data processing
pandas>=2.0.0

# PDF processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Image/OCR processing
pytesseract>=0.3.10
Pillow>=10.0.0

# Excel processing
openpyxl>=3.1.0

# HTTP client for Mistral OCR
httpx>=0.27.0

# Optional env file loading
python-dotenv>=1.0.0

